server change

+++ routes/web.php
@@ -54,7 +54,7 @@ Route::middleware(['auth', 'isAdmin'])->prefix('admin')->group(function () {
     Route::get('products', [AdminController::class, 'products']);
     Route::get('stores', [AdminController::class, 'stores']);
     Route::delete('user/{id}', [AdminController::class, 'delete_user']);
-    Route::delete('product/{id}', [AdminController::class, 'delete_product']);
+    Route::delete('product/{id}', [AdminController::class, 'delete_product'])->name('product.admin.delete');
     // Route::delete('store{id}', [AdminController::class, 'delete_store']);
     Route::delete('reset_data', [AdminController::class, 'maintenance']);
     Route::resource('initiators', InitiatorController::class);
